requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
motor==3.3.2
packaging==26.0
pluggy==1.6.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.16.0
pytest==8.0.2
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.28.0

